@@ -20,8 +20,8 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 # Read more about Android versioning at https://developer.android.com/studio/publish/versioning
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
-# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 0.0.0+1
+# https://develoyper.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
+version: 1.0.0+2
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
@@ -34,11 +34,11 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.3
   # if you are using latest flutter please update your flutter_svg package to ^0.18.*
-  flutter_svg: ^0.22.0 # help us to use SVG in our app
+  flutter_svg: ^2.0.9 # help us to use SVG in our app
   provider: ^6.0.0 # for State management
   swipe_cards: ^2.0.0+1
-  flutter_secure_storage: ^8.0.0
-  phone_number: ^1.0.0
+  flutter_secure_storage: ^9.0.0
+  phone_number: ^2.0.1
   flutter_dotenv: ^5.1.0
   http: ^0.13.6
   otp_text_field: ^1.1.3
@@ -46,14 +46,14 @@ dependencies:
   string_to_color: ^2.0.0
   mixpanel_flutter: ^2.1.1
   buttons_tabbar: ^1.3.7+1
-  google_maps_flutter: ^2.3.0
+  google_maps_flutter: ^2.5.0
   flutter_google_places: ^0.3.0
   collection: ^1.17.0
   email_validator: ^2.1.17
   flutter_initicon: ^3.0.0+1
   easy_stepper: ^0.6.0
   visibility_detector: ^0.4.0+2
-  flutter_launcher_icons: ^0.11.0
+  flutter_launcher_icons: ^0.13.1
   flutter_web_auth: ^0.5.0
   uni_links: ^0.5.1
   flutter_contacts: ^1.1.7+1
@@ -62,9 +62,9 @@ dependencies:
   webview_flutter: ^4.4.2
   lottie: ^2.3.2
   animated_text_kit: ^4.2.2
-  rounded_background_text: ^0.3.0
-  url_launcher: ^6.1.11
-  font_awesome_flutter: ^10.4.0
+  rounded_background_text: ^0.4.3
+  url_launcher: ^6.2.2
+  font_awesome_flutter: ^10.6.0
   device_info_plus: ^9.1.1
   android_id: ^0.3.6
   uuid: ^4.1.0
