@@ -2,6 +2,7 @@ import 'package:Wishy/components/delivery_availability_dialog.dart';
 import 'package:Wishy/global_manager.dart';
 import 'package:Wishy/screens/checkout/checkout_screen.dart';
 import 'package:Wishy/utils/analytics.dart';
+import 'package:Wishy/utils/is_variants_exists.dart';
 import 'package:flutter/material.dart';
 import 'package:Wishy/components/additional_details_dialog.dart';
 import 'package:Wishy/components/default_button.dart';
@@ -64,7 +65,7 @@ class Body extends StatelessWidget {
               ),
               if (product.variants == null)
                 buildOutOfStock(secondColor)
-              else if (product.variants!.length > 1)
+              else if (isVariantsExists(product.variants))
                 VariantsWidget(
                     situation: situation,
                     productId: product.id,
